backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.5
Django==6.0.3
django-cors-headers==4.9.0
django-otp==1.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
iniconfig==2.3.0
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.12.1
pyotp==2.9.0
pytest==9.0.2
pytest-django==4.12.0
qrcode==8.2
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3
